add: visor in streamlit
