--- src/visor/install/my-test-package/share/my-test-package/hook/pythonpath.dsv
@@ -1,1 +1,0 @@
-prepend-non-duplicate;PYTHONPATH;lib/python3.10/site-packages
