--- src/visor/install/my-test-package/share/my-test-package/hook/pythonpath.ps1
@@ -1,3 +1,0 @@
-# generated from colcon_powershell/shell/template/hook_prepend_value.ps1.em
-
-colcon_prepend_unique_value PYTHONPATH "$env:COLCON_CURRENT_PREFIX\lib/python3.10/site-packages"
