--- src/visor/install/my-test-package/share/my-test-package/hook/pythonpath.sh
@@ -1,3 +1,0 @@
-# generated from colcon_core/shell/template/hook_prepend_value.sh.em
-
-_colcon_prepend_unique_value PYTHONPATH "$COLCON_CURRENT_PREFIX/lib/python3.10/site-packages"
