--- src/visor/install/my-test-package/share/my-test-package/package.bash
@@ -1,31 +1,0 @@
-# generated from colcon_bash/shell/template/package.bash.em
-
-# This script extends the environment for this package.
-
-# a bash script is able to determine its own path if necessary
-if [ -z "$COLCON_CURRENT_PREFIX" ]; then
-  # the prefix is two levels up from the package specific share directory
-  _colcon_package_bash_COLCON_CURRENT_PREFIX="$(builtin cd "`dirname "${BASH_SOURCE[0]}"`/../.." > /dev/null && pwd)"
-else
-  _colcon_package_bash_COLCON_CURRENT_PREFIX="$COLCON_CURRENT_PREFIX"
-fi
-
-# function to source another script with conditional trace output
-# first argument: the path of the script
-# additional arguments: arguments to the script
-_colcon_package_bash_source_script() {
-  if [ -f "$1" ]; then
-    if [ -n "$COLCON_TRACE" ]; then
-      echo "# . \"$1\""
-    fi
-    . "$@"
-  else
-    echo "not found: \"$1\"" 1>&2
-  fi
-}
-
-# source sh script of this package
-_colcon_package_bash_source_script "$_colcon_package_bash_COLCON_CURRENT_PREFIX/share/my-test-package/package.sh"
-
-unset _colcon_package_bash_source_script
-unset _colcon_package_bash_COLCON_CURRENT_PREFIX
