--- src/visor/install/my-test-package/share/my-test-package/package.dsv
@@ -1,3 +1,0 @@
-source;share/my-test-package/hook/pythonpath.ps1
-source;share/my-test-package/hook/pythonpath.dsv
-source;share/my-test-package/hook/pythonpath.sh
