--- src/visor/log/latest
@@ -1,1 +1,0 @@
-latest_build
