--- src/visor/log/latest_build
@@ -1,1 +1,0 @@
-build_2024-05-21_16-22-39
