+++ src/visor_v2/main.py
@@ -3,9 +3,15 @@ from rclpy.node import Node
 from rclpy.qos import qos_profile_sensor_data
 from geometry_msgs.msg import Twist
 from sensor_msgs.msg import LaserScan
+from std_msgs.msg import String
 import streamlit as st
 import threading
 import time
+import cv2
+import base64
+import numpy as np
+from PIL import Image
+import io
 
 class RobotController(Node):
     def __init__(self):
@@ -97,43 +103,88 @@ class RobotController(Node):
     def start_switch(self):
         self.killed = False
 
-def run_ros2_listener(robot_controller):
+class Listener(Node):
+    def __init__(self):
+        super().__init__('listener')
+        self.subscription = self.create_subscription(
+            String,
+            'chatter',
+            self.listener_callback,
+            10)
+        self.subscription  # prevent unused variable warning
+
+        # Configuração do Streamlit
+        st.title("Robot Controller via ROS 2 and Streamlit")
+        self.frame_holder = st.empty()
+        self.latency_placeholder = st.empty()
+
+    def listener_callback(self, msg):
+        timestamp, jpg_as_text = msg.data.split('|', 1)
+        timestamp = float(timestamp)
+        current_time = time.time()
+        latency = current_time - timestamp
+
+        jpg_original = base64.b64decode(jpg_as_text)
+        jpg_as_np = np.frombuffer(jpg_original, dtype=np.uint8)
+        img = cv2.imdecode(jpg_as_np, cv2.IMREAD_COLOR)
+        if img is not None:
+            img_rgb = cv2.cvtColor(img, cv2.COLOR_BGR2RGB)
+            pil_image = Image.fromarray(img_rgb)
+            img_bytes = io.BytesIO()
+            pil_image.save(img_bytes, format="JPEG")
+            img_bytes.seek(0)
+            self.frame_holder.image(
+                img_bytes, caption="Webcam Stream", use_column_width=True
+            )
+            self.latency_placeholder.text(f"Latency: {latency:.4f} seconds")
+        else:
+            self.get_logger().error('Could not decode the image')
+
+def main():
+    rclpy.init()
+    robot_controller = RobotController()
+    listener = Listener()
+
+    def run_ros2_listener():
+        rclpy.spin(listener)
+
+    threading.Thread(target=run_ros2_listener, daemon=True).start()
+
+    st.sidebar.title("Control Panel")
+
+    if st.sidebar.button("Connect"):
+        robot_controller.connect()
+
+    if st.sidebar.button("Disconnect"):
+        robot_controller.disconnect()
+
+    if st.sidebar.button("Increase Linear Speed"):
+        robot_controller.increase_linear_speed()
+
+    if st.sidebar.button("Decrease Linear Speed"):
+        robot_controller.decrease_linear_speed()
+
+    if st.sidebar.button("Increase Angular Speed"):
+        robot_controller.increase_angular_speed()
+
+    if st.sidebar.button("Decrease Angular Speed"):
+        robot_controller.decrease_angular_speed()
+
+    if st.sidebar.button("Stop Robot"):
+        robot_controller.stop_robot()
+
+    if st.sidebar.button("Kill Switch"):
+        robot_controller.kill_switch()
+
+    if st.sidebar.button("Start Switch"):
+        robot_controller.start_switch()
+
+    if st.sidebar.button("Exit"):
+        robot_controller.disconnect()
+        rclpy.shutdown()
+        st.stop()
+
     rclpy.spin(robot_controller)
 
-rclpy.init()
-robot_controller = RobotController()
-threading.Thread(target=run_ros2_listener, args=(robot_controller,), daemon=True).start()
-
-st.title("Robot Controller via ROS 2 and Streamlit")
-
-if st.button("Connect"):
-    robot_controller.connect()
-
-if st.button("Disconnect"):
-    robot_controller.disconnect()
-
-if st.button("Increase Linear Speed"):
-    robot_controller.increase_linear_speed()
-
-if st.button("Decrease Linear Speed"):
-    robot_controller.decrease_linear_speed()
-
-if st.button("Increase Angular Speed"):
-    robot_controller.increase_angular_speed()
-
-if st.button("Decrease Angular Speed"):
-    robot_controller.decrease_angular_speed()
-
-if st.button("Stop Robot"):
-    robot_controller.stop_robot()
-
-if st.button("Kill Switch"):
-    robot_controller.kill_switch()
-
-if st.button("Start Switch"):
-    robot_controller.start_switch()
-
-if st.button("Exit"):
-    robot_controller.disconnect()
-    rclpy.shutdown()
-    st.stop()
+if __name__ == '__main__':
+    main()
